@@ -26,29 +26,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.12",
-    "@angular/common": "13.3.12",
-    "@angular/compiler": "13.3.12",
-    "@angular/core": "13.3.12",
-    "@angular/forms": "13.3.12",
-    "@angular/platform-browser": "13.3.12",
-    "@angular/platform-browser-dynamic": "13.3.12",
-    "@angular/router": "13.3.12",
+    "@angular/animations": "14.2.12",
+    "@angular/common": "14.2.12",
+    "@angular/compiler": "14.2.12",
+    "@angular/core": "14.2.12",
+    "@angular/forms": "14.2.12",
+    "@angular/platform-browser": "14.2.12",
+    "@angular/platform-browser-dynamic": "14.2.12",
+    "@angular/router": "14.2.12",
     "@ngx-translate/core": "14.0.0",
     "@ngx-translate/http-loader": "7.0.0",
     "ng-connection-service": "9.0.3",
-    "ngx-infinite-scroll": "13.1.0",
-    "ngx-toastr": "14.3.0",
+    "ngx-infinite-scroll": "14.0.1",
+    "ngx-toastr": "15.2.2",
     "ramda": "0.27.2",
     "rxjs": "6.5.5",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.10",
-    "@angular/cli": "13.3.10",
-    "@angular/compiler-cli": "13.3.12",
-    "@angular/language-service": "13.3.12",
+    "@angular-devkit/build-angular": "14.2.10",
+    "@angular/cli": "14.2.10",
+    "@angular/compiler-cli": "14.2.12",
+    "@angular/language-service": "14.2.12",
     "@biesbjerg/ngx-translate-extract": "7.0.4",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "2.0.8",
@@ -66,7 +66,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.2.2",
     "protractor": "~7.0.0",
     "rxjs-tslint": "0.1.8",
     "ts-node": "8.8.2",
